03-chicago-transit/requirements.txt
# ─── Core ────────────────────────────────────────────────────────────────────
python-dotenv==1.0.0
requests==2.31.0
jsonschema==4.21.1

# ─── AWS ─────────────────────────────────────────────────────────────────────
boto3==1.34.34

# ─── GCP ─────────────────────────────────────────────────────────────────────
google-cloud-pubsub==2.20.1
google-cloud-bigquery==3.17.1
google-cloud-bigquery-storage==2.24.0
google-auth==2.28.0
db-dtypes==1.2.0

# ─── Orchestration ───────────────────────────────────────────────────────────
prefect==2.16.4

# ─── Data processing ─────────────────────────────────────────────────────────
pandas==2.2.0
numpy==1.26.4
pyarrow==15.0.0

# ─── Forecasting ─────────────────────────────────────────────────────────────
prophet==1.1.5
holidays==0.43

# ─── Statistical analysis ────────────────────────────────────────────────────
statsmodels==0.14.1
scipy==1.12.0

# ─── Visualization (notebooks) ───────────────────────────────────────────────
matplotlib==3.8.2
seaborn==0.13.2
plotly==5.19.0

# ─── dbt ─────────────────────────────────────────────────────────────────────
dbt-bigquery==1.7.4
dbt-core==1.7.4

# ─── Testing ─────────────────────────────────────────────────────────────────
pytest==8.0.0
pytest-cov==4.1.0
responses==0.25.0

# ─── Notebooks ───────────────────────────────────────────────────────────────
jupyterlab==4.1.1
ipykernel==6.29.2
